zephyr: ll-schedule: restore clock after idle in scheduler
The adsp_clock_idle_exit function is invoked at the beginning of the
scheduler thread function to reset the clock frequency if it was lowered
during idle periods. This change is crucial for maintaining the
performance and responsiveness of the system when it resumes work after
being idle.

This patch is conditionally compiled with
CONFIG_ADSP_DYNAMIC_CLOCK_SWITCHING to ensure that the functionality is
available only on systems that support dynamic clock management.

Signed-off-by: Tomasz Leman <tomasz.m.leman@intel.com>
(cherry picked from commit af791dee0807ae9227bf94d252c4ed6358cba662)